refactor(checkpoint): move the VM-state probes out, and exempt them where they landed
Two follow-ons the enforcement commit needs to stand on its own now that A1/A2
are on main.

The enforcement work takes checkpoint.rs past the 1500-line production ceiling
again (1526), so the VM-state probes move to `checkpoint/vm_state.rs`: ten
helpers answering one question between them — is this VM quiesced enough to
checkpoint, where is its rootfs, can its backend save and restore. 1526 -> 1383.

Moving them broke `check-cli-runtime-surface`, whose exemptions are keyed by
file path, so `backend_for_vm` and `ensure_save_restore_supported` lost the
entry that covered them in checkpoint.rs. Added one for the new path rather than
a directory prefix — the neighbouring entry says why: a prefix would silently
exempt whatever lands beside it next.

Three test call sites that came from main predate `allow_secret_drop` and now
set it explicitly to `false`, the strict default: they do not exercise
attenuation, and defaulting to permissive in a test is how a refusal stops being
tested.
